chore: move extension to projects
