+++ projects/UnrealBridge/handler/unrealmisc.h
