+++ projects/UnrealBridge/initbridge.cpp
@@ -1,0 +1,13 @@
+#include <zeno/core/Session.h>
+#include <zeno/extra/EventCallbacks.h>
+#include "unrealtcpserver.h"
+#include "unrealudpserver.h"
+
+namespace zeno {
+
+static int defUnrealBridgeInit = getSession().eventCallbacks->hookEvent("init", [] {
+    zeno::startUnrealTcpServer();
+    zeno::startUnrealUdpServer();
+});
+
+}
