+++ ui/zenoedit/launch/viewdecode.cpp
@@ -10,7 +10,9 @@
 #include <zeno/extra/GlobalComm.h>
 #include <zeno/extra/GlobalStatus.h>
 #include <zeno/funcs/ObjectCodec.h>
-#include "unreal/unrealhook.h"
+#ifdef ZENO_WITH_UnrealBridge
+#include "unrealhook.h"
+#endif
 #include <rapidjson/document.h>
 #include <type_traits>
 #include <iostream>
@@ -76,7 +78,9 @@ struct PacketProc {
                 zeno::log_warn("failed to decode view object");
                 return false;
             }
+#ifdef ZENO_WITH_UnrealBridge
             zeno::UnrealHook::fetchViewObject(objKey, object);
+#endif
             clearGlobalIfNeeded();
             zeno::getSession().globalComm->addViewObject(objKey, object);
